requirements.txt
Flask==2.3.3
psutil==5.9.5
pyTelegramBotAPI==4.14.0
speedtest-cli==2.1.3
schedule==1.2.0
python-dotenv==1.0.1
ntplib==0.4.0
